+++ CHANGELOG.md
@@ -11,4 +11,5 @@ PostHog JS is a mono-repo. The changelogs are in the individual packages
 - [@posthog/convex](./packages/convex/CHANGELOG.md)
 - [@posthog/next](./packages/next/CHANGELOG.md)
 - [@posthog/nextjs-config](./packages/nextjs-config/CHANGELOG.md)
+- [@posthog/openfeature-node-provider](./packages/openfeature-node-provider/CHANGELOG.md)
 - [@posthog/plugin-utils](./packages/plugin-utils/CHANGELOG.md)